added audio/visual, new model for ship, added teleport, and win volume

+++ Assets/Scripts/hazardVolume.cs
@@ -4,13 +4,18 @@ using UnityEngine;
 
 public class hazardVolume : MonoBehaviour
 {
+    [Header("Audio")]
+    [SerializeField] AudioClip _sound = null;
+    [SerializeField] float _volume = 0;
+
     private void OnTriggerEnter(Collider other)
     {
         PlayerShip playShip = other.gameObject.GetComponent<PlayerShip>();
 
-        if(playShip != null)
+        if(playShip != null && other.CompareTag("Player"))
         {
             playShip.kill();
+            AudioHelper.PlayClip2D(_sound, _volume);
         }
     }
 }
